Separate core.streaming
