+++ src/main/kotlin/jp/nephy/penicillin/core/request/action/StreamApiAction.kt
@@ -3,8 +3,8 @@ package jp.nephy.penicillin.core.request.action
 import jp.nephy.penicillin.core.exceptions.PenicillinException
 import jp.nephy.penicillin.core.request.ApiRequest
 import jp.nephy.penicillin.core.response.StreamResponse
-import jp.nephy.penicillin.core.streaming.StreamHandler
-import jp.nephy.penicillin.core.streaming.StreamListener
+import jp.nephy.penicillin.core.streaming.handler.StreamHandler
+import jp.nephy.penicillin.core.streaming.listener.StreamListener
 import kotlinx.coroutines.CancellationException
 
 data class StreamApiAction<L: StreamListener, H: StreamHandler<L>>(override val request: ApiRequest): ApiAction<StreamResponse<L, H>> {
