+++ src/main/kotlin/jp/nephy/penicillin/core/streaming/handler/FilterStreamHandler.kt
@@ -1,12 +1,13 @@
-package jp.nephy.penicillin.core.streaming
+package jp.nephy.penicillin.core.streaming.handler
 
 import jp.nephy.jsonkt.JsonObject
+import jp.nephy.penicillin.core.streaming.listener.FilterStreamListener
 import jp.nephy.penicillin.models.Status
 import jp.nephy.penicillin.models.Stream
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
-class SampleStreamHandler(override val listener: SampleStreamListener): StreamHandler<SampleStreamListener> {
+class FilterStreamHandler(override val listener: FilterStreamListener): StreamHandler<FilterStreamListener> {
     override suspend fun handle(json: JsonObject, scope: CoroutineScope) {
         scope.launch(scope.coroutineContext) {
             when {
@@ -24,9 +25,4 @@ class SampleStreamHandler(override val listener: SampleStreamListener): StreamHa
 
         listener.onAnyJson(json)
     }
-}
-
-interface SampleStreamListener: StreamListener {
-    suspend fun onStatus(status: Status) {}
-    suspend fun onDelete(delete: Stream.Delete) {}
 }
