+++ src/main/kotlin/jp/nephy/penicillin/core/streaming/handler/SampleStreamHandler.kt
@@ -1,12 +1,13 @@
-package jp.nephy.penicillin.core.streaming
+package jp.nephy.penicillin.core.streaming.handler
 
 import jp.nephy.jsonkt.JsonObject
+import jp.nephy.penicillin.core.streaming.listener.SampleStreamListener
 import jp.nephy.penicillin.models.Status
 import jp.nephy.penicillin.models.Stream
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 
-class FilterStreamHandler(override val listener: FilterStreamListener): StreamHandler<FilterStreamListener> {
+class SampleStreamHandler(override val listener: SampleStreamListener): StreamHandler<SampleStreamListener> {
     override suspend fun handle(json: JsonObject, scope: CoroutineScope) {
         scope.launch(scope.coroutineContext) {
             when {
@@ -24,9 +25,4 @@ class FilterStreamHandler(override val listener: FilterStreamListener): StreamHa
 
         listener.onAnyJson(json)
     }
-}
-
-interface FilterStreamListener: StreamListener {
-    suspend fun onStatus(status: Status) {}
-    suspend fun onDelete(delete: Stream.Delete) {}
 }
