+++ src/main/kotlin/jp/nephy/penicillin/core/streaming/handler/StreamHandler.kt
@@ -1,6 +1,7 @@
-package jp.nephy.penicillin.core.streaming
+package jp.nephy.penicillin.core.streaming.handler
 
 import jp.nephy.jsonkt.JsonObject
+import jp.nephy.penicillin.core.streaming.listener.StreamListener
 import kotlinx.coroutines.CoroutineScope
 
 interface StreamHandler<L: StreamListener> {
