+++ src/main/kotlin/jp/nephy/penicillin/endpoints/LivePipeline.kt
@@ -3,8 +3,8 @@
 package jp.nephy.penicillin.endpoints
 
 import jp.nephy.penicillin.PenicillinClient
-import jp.nephy.penicillin.core.streaming.LivePipelineHandler
-import jp.nephy.penicillin.core.streaming.LivePipelineListener
+import jp.nephy.penicillin.core.streaming.handler.LivePipelineHandler
+import jp.nephy.penicillin.core.streaming.listener.LivePipelineListener
 import jp.nephy.penicillin.models.LivePipelineSubscription
 
 class LivePipeline(override val client: PenicillinClient): Endpoint {
